Today I animated the gradient on the background. I found this really helpful guide https://www.geeksforgeeks.org/how-to-animate-gradient-background-smoothly-using-css/. This thought me a lot about CSS animation.
I discovered that I could do these animations using keyframes. I learned that Keyframes are essentially various snapshots of how the animation and the stages of the animation.
I decide to use the built-in animation in CSS called
ease infinite.

 To animate the gradient, I defined a CSS animation using keyframes. In this case, I created an animation called background gradient.

At first, I thought I just had to type in the name of the animation and set a time. But then I realised when it never worked that I also need to add keyframes.
In the key frames of my animation, I targeted the backgorund gradient animation.  The 0% indicates the animation's beginning, the 50% is its middle, and the 100% is its conclusion. Each keyframe contains a background position that instructs the browser where to place the background picture at that precise moment in time.

I set the animation to 15s  an ease timing function for smooth transitions and set it to repeat infinite times to create a continuous animation loop.