oday, I fixed the formatting of the "Back to Home" or "Go to Level Menu" button that appears at the end of the quiz. I also added a percentage for the user's score, which is displayed at the end of the quiz. Additionally, I reset the points to 0 and the question number if the user wants to replay the quiz. To improve code organization, I added some more divs.
For the end score report, instead of modifying the existing HTML, I created a new div so that I can hide and unhide the divs as needed. This approach is better practice. Instead of changing existing HTML and adding new HTML elements like percentage text and the final score message, I used JavaScript to update the inner html property. I also worked on hiding divs when they are not required to be displayed, such as the score report when the user clicks "Play Again."